Rename example sketch file to match folder name
The Arduino IDE requires the sketch folder name to match the filename of the primary sketch file. On a case-sensitive OS like Linux this includes filename case matching. This change causes the example sketch to be accessible via the Arduino IDE's File > Examples > LIBRARYNAME menu after the library is installed.

Reference:
https://github.com/arduino/Arduino/wiki/Arduino-IDE-1.5:-Library-specification#library-examples